requirements.txt
python-dotenv==1.0.1
slack-bolt==1.18.1
slack_sdk==3.27.1